cmd/forwarder: move version command to separate package
